# 5x Faster `IntersectionType->getEnumCases()`