--- a/src/Type/IntersectionType.php
+++ b/src/Type/IntersectionType.php
@@ -842,7 +842,7 @@ class IntersectionType implements CompoundType
 		foreach ($this->types as $type) {
 			$oneType = [];
 			foreach ($type->getEnumCases() as $enumCase) {
-				$oneType[md5($enumCase->describe(VerbosityLevel::typeOnly()))] = $enumCase;
+				$oneType[$enumCase->getClassName() . '::' . $enumCase->getEnumCaseName()] = $enumCase;
 			}
 			$compare[] = $oneType;
 		}